fix bug

+++ src/DataTypes.java
@@ -1,9 +1,9 @@
 import java.util.List;
 
 public class DataTypes {
-    public static int sum(List<Integer> numbers) {
+    public static long sum(List<Integer> numbers) {
 
-        int s = 0;
+        long s = 0;
         // below is a "foreach" loop which iterates through numbers
         for (int x : numbers) {
             s += x;
